src/pyproject.toml
[tool.poetry]
name = "work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gspread = "^5.12.2"
oauth2client = "^4.1.3"
paramiko = "^3.4.0"
tenacity = "^9.0.0"
orgparse = "^0.4.20231004"
google-api-python-client = "^2.166.0"
google-auth = "^2.38.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
pytest = "^8.0.2"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
